+++ CHANGELOG.md
@@ -7,4 +7,7 @@ CHANGELOG
  * create tests 
 
 
- 
+0.2 2022-02-16
+-----
+ * `PropertyIndex` class extends new `PropertyPicker` class 
+ * create new `PropertyTreeBuilder` class